Refactor send viewControllers
